refactor: add unit-postdetail module
