+++ core-commonui/src/iosMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/core/commonui/di/Utils.kt
@@ -1,22 +1,12 @@
 package com.github.diegoberaldin.raccoonforlemmy.core.commonui.di
 
 import com.github.diegoberaldin.raccoonforlemmy.core.commonui.communitydetail.CommunityDetailMviModel
-import com.github.diegoberaldin.raccoonforlemmy.core.commonui.postdetail.PostDetailMviModel
 import com.github.diegoberaldin.raccoonforlemmy.core.commonui.userdetail.UserDetailMviModel
 import com.github.diegoberaldin.raccoonforlemmy.domain.lemmy.data.CommunityModel
-import com.github.diegoberaldin.raccoonforlemmy.domain.lemmy.data.PostModel
 import com.github.diegoberaldin.raccoonforlemmy.domain.lemmy.data.UserModel
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
 import org.koin.core.parameter.parametersOf
-
-actual fun getPostDetailViewModel(
-    post: PostModel,
-    otherInstance: String,
-    highlightCommentId: Int?,
-    isModerator: Boolean,
-): PostDetailMviModel =
-    CommonUiViewModelHelper.getPostDetailModel(post, otherInstance, highlightCommentId, isModerator)
 
 actual fun getCommunityDetailViewModel(
     community: CommunityModel,
@@ -28,19 +18,6 @@ actual fun getUserDetailViewModel(user: UserModel, otherInstance: String): UserD
     CommonUiViewModelHelper.getUserDetailModel(user, otherInstance)
 
 object CommonUiViewModelHelper : KoinComponent {
-
-    fun getPostDetailModel(
-        post: PostModel,
-        otherInstance: String,
-        highlightCommentId: Int?,
-        isModerator: Boolean,
-    ): PostDetailMviModel {
-        val model: PostDetailMviModel by inject(
-            parameters = { parametersOf(post, otherInstance, highlightCommentId, isModerator) },
-        )
-        return model
-    }
-
     fun getCommunityDetailModel(
         community: CommunityModel,
         otherInstance: String,
