+++ core-commonui/unit-postdetail/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/unit/postdetail/di/Utils.kt
@@ -1,0 +1,11 @@
+package com.github.diegoberaldin.raccoonforlemmy.unit.postdetail.di
+
+import com.github.diegoberaldin.raccoonforlemmy.domain.lemmy.data.PostModel
+import com.github.diegoberaldin.raccoonforlemmy.unit.postdetail.PostDetailMviModel
+
+expect fun getPostDetailViewModel(
+    post: PostModel,
+    otherInstance: String = "",
+    highlightCommentId: Int? = null,
+    isModerator: Boolean = false,
+): PostDetailMviModel
