+++ settings.gradle.kts
@@ -56,5 +56,6 @@ include(":core-commonui:unit-saveditems")
 include(":core-commonui:unit-createreport")
 include(":core-commonui:unit-createcomment")
 include(":core-commonui:unit-createpost")
+include(":core-commonui:unit-postdetail")
 include(":core-commonui:detailopener-api")
 include(":core-commonui:detailopener-impl")
